Refactor westures-core access (simplify)
